+++ toolium/pageelements/group_page_element.py
@@ -24,7 +24,8 @@ from toolium.pageobjects.page_object import PageObject
 
 
 class Group(PageObject, PageElement):
-    def __init__(self, by, value, parent=None, driver_wrapper=None, order=None, wait=False, webview=False):
+    def __init__(self, by, value, parent=None, driver_wrapper=None, order=None, wait=False, webview=False,
+                 webview_context_selection_callback=None, webview_csc_args=None):
         """Initialize the Group object with the given locator components.
 
         If parent is not None, find_elements will be performed over it, instead of
@@ -38,6 +39,10 @@ class Group(PageObject, PageElement):
         :param wait: True if the page element must be loaded in wait_until_loaded method of the container page object
         :param shadowroot: CSS SELECTOR of JS element where shadowroot tag appears
         :param webview: True if the element is in a mobile webiew
+        :param webview_context_selection_callback: method provided to select the desired webview context if
+        automatic_context_selection is enabled. Must return a tuple (context, window_handle) for android, and a context
+        for ios
+        :param webview_csc_args: arguments list for webview_context_selection_callback
         """
         self.logger = logging.getLogger(__name__)  #: logger instance
         self.locator = (by, value)  #: tuple with locator type and locator value
@@ -45,6 +50,9 @@ class Group(PageObject, PageElement):
         self.order = order  #: index value if the locator returns more than one element
         self.wait = wait  #: True if it must be loaded in wait_until_loaded method of the container page object
         self.webview = webview  #: True if element is in a mobile webview
+        self.webview_context_selection_callback = webview_context_selection_callback  #: callback for selection of the
+        # webview context with automatic_context_selection
+        self.webview_csc_args = webview_csc_args  #: arguments list for the context selection callback method
         self.shadowroot = None  #: Not implemented for Group yet
         self.driver_wrapper = driver_wrapper if driver_wrapper else \
             DriverWrappersPool.get_default_wrapper()  #: driver wrapper instance