+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ v1.8.3
 - Added utilities to download files
 - Get text for InputText element in mobile tests
 - Add dataset utilities
+- Manage multiples webviews for mobile tests
 
 v1.8.2
 ------